+++ src/errors.rs
@@ -1,3 +1,5 @@
+use bevy::ecs::error::BevyError;
+
 #[derive(thiserror::Error, Debug)]
 pub(crate) enum YoleckAssetLoaderError {
     #[error("{0}")]
@@ -7,5 +9,12 @@ pub(crate) enum YoleckAssetLoaderError {
     #[error("{0}")]
     SerdeJson(#[from] serde_json::Error),
     #[error("{0}")]
-    Anyhow(#[from] anyhow::Error),
+    Bevy(BevyError),
+}
+
+// For some reason #[from] doesn't work...
+impl From<BevyError> for YoleckAssetLoaderError {
+    fn from(value: BevyError) -> Self {
+        Self::Bevy(value)
+    }
 }